# swagger.yaml
openapi: 3.1.0
info:
  title: AI OCR Service
  version: 0.1.0
paths:
  /ocr/extract:
    post:
      tags:
      - OCR
      summary: Extract Text
      description: Extracts text from a PDF file using Typhoon OCR.
      operationId: extract_text_ocr_extract_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_extract_text_ocr_extract_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sheets/analyze:
    post:
      tags:
      - Sheets
      summary: Analyze Sheet
      description: 'Async AI Analysis Pipeline:

        1. Upload File (PDF) OR Provide URL

        2. Create Job (Pending)

        3. Return Job ID (202 Accepted)

        4. Background: OCR -> AI -> DB Update -> Webhook'
      operationId: analyze_sheet_sheets_analyze_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_analyze_sheet_sheets_analyze_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sheets/jobs/{job_id}:
    get:
      tags:
      - Sheets
      summary: Get Job Status
      operationId: get_job_status_sheets_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sheets/jobs/by-sheet/{sheet_id}:
    get:
      tags:
      - Sheets
      summary: Get Job By Sheet
      description: 'Poll job status by sheet_id.

        Returns the latest job associated with this sheet_id.'
      operationId: get_job_by_sheet_sheets_jobs_by_sheet__sheet_id__get
      parameters:
      - name: sheet_id
        in: path
        required: true
        schema:
          type: string
          title: Sheet Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/audio/transcribe:
    post:
      tags:
      - Audio
      summary: Transcribe Audio Background
      description: 'Upload an audio file (mp3/wav) for background processing.

        Returns a job_id immediately.'
      operationId: transcribe_audio_background_api_audio_transcribe_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_transcribe_audio_background_api_audio_transcribe_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/chat/:
    post:
      tags:
      - Chat
      summary: Chat Endpoint
      description: 'Chat endpoint supporting RAG context injection and history.

        Returns a unified JSON response.'
      operationId: chat_endpoint_api_chat__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /:
    get:
      summary: Health Check
      operationId: health_check__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    Body_analyze_sheet_sheets_analyze_post:
      properties:
        file:
          anyOf:
          - type: string
            format: binary
          - type: 'null'
          title: File
        file_url:
          anyOf:
          - type: string
          - type: 'null'
          title: File Url
        webhook_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Webhook Url
        sheet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Sheet Id
      type: object
      title: Body_analyze_sheet_sheets_analyze_post
    Body_extract_text_ocr_extract_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_extract_text_ocr_extract_post
    Body_transcribe_audio_background_api_audio_transcribe_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_transcribe_audio_background_api_audio_transcribe_post
    ChatRequest:
      properties:
        session_id:
          type: string
          title: Session Id
          description: Unique session ID for the chat
        message:
          type: string
          title: Message
          description: The user's message
        sheet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Sheet Id
          description: Optional sheet ID for context injection (RAG)
      type: object
      required:
      - session_id
      - message
      title: ChatRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError